[Move #359] "useBeforeUnloadHandler" 경로 변경
