+++ hooks/useBeforeUnloadHandler.ts
